add main.c

+++ sdl.sh
@@ -11,9 +11,17 @@ ls $FRAMEWORKS_DIR
 
 mkdir -p tmp
 SDL2_DMG=SDL2-2.0.8.dmg
-curl -L -o tmp/$SDL2_DMG https://www.libsdl.org/release/$SDL2_DMG
-hdiutil attach tmp/$SDL2_DMG
-cp -R /Volumes/SDL2/$SDL2_FRAMEWORK $FRAMEWORKS_DIR
-hdiutil detach /Volumes/SDL2
+
+if [ ! -e $FRAMEWORKS_DIR/$SDL2_FRAMEWORK ]; then
+  if [ ! -e tmp/$SDL2_DMG ]; then
+    curl -L -o tmp/$SDL2_DMG https://www.libsdl.org/release/$SDL2_DMG
+  fi
+  hdiutil attach tmp/$SDL2_DMG
+  cp -R /Volumes/SDL2/$SDL2_FRAMEWORK $FRAMEWORKS_DIR
+  hdiutil detach /Volumes/SDL2
+fi
 ls tmp/
 ls $FRAMEWORKS_DIR
+
+gcc main.c -framework SDL2 -F ~/Library/Frameworks
+./a.out
